requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
chromadb
pypdf
python-dotenv
sentence-transformers
google-generativeai